Landing Pages: Template Library - Added a Category filter to the Landing Pages tab in the Template Library modal. | Fix - Viewing new Landing Page redirects to the wrong permalink. | Fix - Landing Page is duplicated in the Finder. | Fix - In WP Dashboard Landing Pages page, update 'landing-page' type to 'Landing Page'.
