+++ includes/editor-templates/templates.php
@@ -60,7 +60,7 @@ if ( ! defined( 'ABSPATH' ) ) {
 			var activeType = elementor.templates.getFilter('type');
 			#>
 			<div id="elementor-template-library-filter-toolbar-remote" class="elementor-template-library-filter-toolbar">
-				<# if ( 'page' === activeType || 'landing-page' === activeType ) { #>
+				<# if ( 'page' === activeType ) { #>
 					<div id="elementor-template-library-order">
 						<input type="radio" id="elementor-template-library-order-new" class="elementor-template-library-order-input" name="elementor-template-library-order" value="date">
 						<label for="elementor-template-library-order-new" class="elementor-template-library-order-label"><?php echo __( 'New', 'elementor' ); ?></label>
